Fix reload and retest buttons in submissions details page
