+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/details/SubmissionDetails.module.scss
@@ -83,3 +83,7 @@
   justify-content: space-between;
   margin-bottom: 24px;
 }
+
+.buttonsSection {
+  margin-top: 20px;
+}
