+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/submissions-list/RefreshableSubmissionsList.tsx
@@ -1,8 +1,4 @@
-import React, { useCallback } from 'react';
-import isNil from 'lodash/isNil';
-
-import { useSubmissionsDetails } from '../../../hooks/submissions/use-submissions-details';
-import Button, { ButtonType } from '../../guidelines/buttons/Button';
+import React from 'react';
 
 import SubmissionsList, { ISubmissionsListProps } from './SubmissionsList';
 
@@ -12,40 +8,12 @@ const RefreshableSubmissionsList = ({
     items,
     selectedSubmission,
     className = '',
-}: IRefreshableSubmissionsListProps) => {
-    const {
-        state: { currentSubmission },
-        actions: { getSubmissionResults },
-    } = useSubmissionsDetails();
-
-    const submissionsReloadBtnClassName = 'submissionReloadBtn';
-
-    const handleReloadClick = useCallback(async () => {
-        if (isNil(currentSubmission)) {
-            return;
-        }
-
-        const { problem: { id: problemId }, isOfficial, user: { id: userId } } = currentSubmission;
-
-        await getSubmissionResults(problemId, isOfficial, userId);
-    }, [ currentSubmission, getSubmissionResults ]);
-
-    return (
-        <>
-            <SubmissionsList
-              items={items}
-              selectedSubmission={selectedSubmission}
-              className={className}
-            />
-
-            <Button
-              onClick={handleReloadClick}
-              text="Reload"
-              type={ButtonType.secondary}
-              className={submissionsReloadBtnClassName}
-            />
-        </>
-    );
-};
+}: IRefreshableSubmissionsListProps) => (
+    <SubmissionsList
+      items={items}
+      selectedSubmission={selectedSubmission}
+      className={className}
+    />
+);
 
 export default RefreshableSubmissionsList;
